add node-net subcommand
